refactor(test): tag the e2e/live suite and add the CI lane that runs it (#1760)
Step 3 of #1760. Steps 1 (#1791) and 2 (#1877) are done.

36 test files matching `_(e2e|integration|live)_test.go` now carry `//go:build
e2e`: 35 `_e2e_test.go` plus 1 `_live_test.go`, zero `_integration_test.go`, in
internal/cli (33), internal/workflow (2) and internal/daemon (1). The issue body
says 44 files in internal/cli; re-measured at 6f9552bd it is 36 across three
packages, so scoping this to internal/cli would have left three behind.

THE CI LANE SHIPS IN THE SAME COMMIT, because a tag with no runner stops the
tests forever and a green suite is exactly what that looks like:
  - build-test (unchanged command) now runs the UNTAGGED tests only.
  - a new `e2e (tagged)` job runs `go test -tags e2e ./...`. It uses ./... rather
    than the three packages that hold tagged files today, because a package list
    would silently stop covering a tagged file added to a fourth package later -
    the same silent-stop one level up.
  - race-build now compiles with `-tags e2e`. Without it that compile drops the
    36 files from the race binary AND from the shard plan derived by -test.list,
    with nothing failing. Measured: internal/workflow race list is 1017 untagged
    vs 1022 with the tag vs 1022 at base; internal/daemon 181 / 182 / 182. Race
    coverage is therefore what it was before the tag.

FOUR CATEGORIES OF SHARED TEST HELPER HAD TO MOVE OUT OF THE TAGGED FILES, or the
DEFAULT build stops compiling: 20 symbols across 11 files were referenced by
untagged tests, and types moved with their methods. They now live in
`*_support_test.go` siblings with no build tag.

THE ONE THAT MATTERED MOST WAS TestMain. internal/cli's only TestMain lived in
daemon_flock_singleton_e2e_test.go. It dispatches the hidden `sandbox-exec` shim
and the `daemon run` child re-exec BEFORE m.Run; under `go test` the current
executable IS cli.test, so without that dispatch a test which re-execs itself
restarts the whole package suite recursively. Measured, not reasoned:
TestRuntimeCredentialCurationForegroundAndDaemonE2E runs in 1.460s at base, HUNG
past a 360s timeout with TestMain behind the tag, and runs in 1.559s with TestMain
moved back to the default build. My enumeration of shared symbols had missed it
because the filter skipped every name starting with "Test" - the filter excluded
the one symbol that mattered.

TWO-ARM PROOF, since "the suite is green" is what a silently-skipped set looks
like:
  discovery, `go test -list`:      default 4622, with -tags e2e 4756, delta 134
    internal/cli 1892 / 2020, internal/workflow 1017 / 1022, internal/daemon
    181 / 182 - and 6f9552bd's default discovery is 4756, equal to the tagged arm
    here, so nothing was lost, only moved behind the tag.
  execution, `go test ./...`:      default 319s wall, 39 packages ok
  execution, `go test -tags e2e`:  398s wall, 39 packages ok
Both arms report the SAME single failure, TestApplyReviewPolicyEmptyHomeIsOff
(#1859, ambient ~/.gitmoot config), which also fails identically at 6f9552bd:
366s wall, 39 ok, same one failure. It is not fixed and not tagged away here.

test_binary_test.go stays UNTAGGED deliberately. sharedGitmootTestBinary has
three callers and only one is in the tagged set: pipeline_defaults_test.go is a
plain unit test and sandbox_produce_hook_linux_test.go is platform-constrained
(`//go:build linux`). Tagging the helper would have moved both out of the default
run, which is a bigger change than this step is allowed to make; the default run
therefore still pays the ./cmd/gitmoot build when an untagged caller needs it.

Composition with the existing platform constraints was checked and is not needed
yet: the four test files carrying `//go:build linux`/`unix` are disjoint from the
36, so no file needs `linux && e2e` today.
