+++ internal/cli/daemon_flock_singleton_e2e_test.go
@@ -1,9 +1,10 @@
+//go:build e2e
+
 package cli
 
 import (
 	"bytes"
 	"errors"
-	"fmt"
 	"os"
 	"os/exec"
 	"strconv"
@@ -13,33 +14,6 @@ import (
 
 	"github.com/gitmoot/gitmoot/internal/config"
 )
-
-// daemonRunChildHomeEnv, when set, flips the re-exec'd test binary into a REAL
-// `gitmoot daemon run --home <value>` process (see TestMain). flock(2) is
-// process-scoped — two goroutines cannot exercise it — so the #556 singleton
-// E2E must launch genuine child processes, and re-exec'ing the already-built
-// test binary is how it gets them without shelling out to `go build`.
-const daemonRunChildHomeEnv = "GITMOOT_TEST_DAEMON_RUN_CHILD_HOME"
-
-func TestMain(m *testing.M) {
-	if len(os.Args) > 1 && os.Args[1] == "sandbox-exec" {
-		// Review jobs wrap every runtime through the current executable. Under
-		// go test that executable is cli.test, so dispatch the hidden shim before
-		// m.Run instead of recursively starting the package suite.
-		os.Exit(Run(os.Args[1:], os.Stdout, os.Stderr))
-	}
-	if home := os.Getenv(daemonRunChildHomeEnv); home != "" {
-		os.Exit(Run([]string{"daemon", "run", "--home", home}, os.Stdout, os.Stderr))
-	}
-	code := m.Run()
-	if err := cleanupSharedGitmootTestBinary(); err != nil {
-		fmt.Fprintf(os.Stderr, "clean shared gitmoot test binary: %v\n", err)
-		if code == 0 {
-			code = 1
-		}
-	}
-	os.Exit(code)
-}
 
 // daemonRunProc is one real `daemon run` child process launched by the E2E.
 // stdout/stderr buffers must only be read AFTER done has delivered (cmd.Wait
