+++ internal/cli/execbackend_e2e_test.go
@@ -1,3 +1,5 @@
+//go:build e2e
+
 package cli
 
 import (
@@ -48,31 +50,6 @@ var execBackendLocalEventKindBaseline = []string{
 	"advance_started",
 	"delegation_worktree_removed",
 	"advance_completed",
-}
-
-var errP2ProbeSubprocessReached = errors.New("p2-probe subprocess runner reached")
-
-type p2ProbeSubprocessRunner struct {
-	calls int
-}
-
-func (r *p2ProbeSubprocessRunner) Run(context.Context, string, string, ...string) (subprocess.Result, error) {
-	r.calls++
-	return subprocess.Result{}, errP2ProbeSubprocessReached
-}
-
-func (r *p2ProbeSubprocessRunner) RunEnv(ctx context.Context, dir string, _ []string, command string, args ...string) (subprocess.Result, error) {
-	return r.Run(ctx, dir, command, args...)
-}
-
-func (r *p2ProbeSubprocessRunner) RunExactEnv(ctx context.Context, dir string, _ []string, _, _ io.Writer, command string, args ...string) error {
-	_, err := r.Run(ctx, dir, command, args...)
-	return err
-}
-
-func (r *p2ProbeSubprocessRunner) LookPath(string) (string, error) {
-	r.calls++
-	return "", errP2ProbeSubprocessReached
 }
 
 // execBackendDispatchAsk enqueues a background shell ask and returns its job id.
