+++ internal/cli/pipeline_run_e2e_test.go
@@ -1,3 +1,5 @@
+//go:build e2e
+
 package cli
 
 import (
@@ -11,32 +13,6 @@ import (
 	"github.com/gitmoot/gitmoot/internal/pipeline"
 	"github.com/gitmoot/gitmoot/internal/runtime"
 )
-
-// pipelineStageResultCmd is a shell command that ignores its input and echoes a
-// valid gitmoot_result with the given decision (and, for a block, the given
-// needs). It is the SHELL-runtime session body a stage job runs as
-// `sh -c <cmd> gitmoot <prompt>`, so the whole pipeline chain runs with NO LLM and
-// NO network — fully deterministic offline (the #446/#533 shell-runtime idiom).
-func pipelineStageResultCmd(decision, summary string, needs []string) string {
-	return pipelineStageResultCmdWithSeverity(decision, summary, needs, "")
-}
-
-func pipelineStageResultCmdWithSeverity(decision, summary string, needs []string, severity string) string {
-	needsJSON := "[]"
-	if len(needs) > 0 {
-		quoted := make([]string, 0, len(needs))
-		for _, n := range needs {
-			quoted = append(quoted, `"`+n+`"`)
-		}
-		needsJSON = "[" + strings.Join(quoted, ",") + "]"
-	}
-	severityJSON := ""
-	if severity != "" {
-		severityJSON = `,"severity":"` + severity + `"`
-	}
-	return `printf '%s' '{"gitmoot_result":{"decision":"` + decision + `"` + severityJSON + `,"summary":"` + summary +
-		`","findings":[],"changes_made":[],"tests_run":[],"needs":` + needsJSON + `,"delegations":[]}}'`
-}
 
 // TestPipelineBlockedParkE2E is the full-chain, NO-LLM, deterministic blocked-park
 // E2E (#681). It drives the REAL chain a daemon iteration runs, end to end:
@@ -306,14 +282,4 @@ stages:
 	if count != 1 {
 		t.Fatalf("advance_skipped_no_pr count = %d, want 1; events=%+v", count, events)
 	}
-}
-
-// pipelineE2EStage renders one stage block with the shell cmd as a literal YAML
-// block scalar (so the JSON-bearing single-quoted command survives YAML parsing).
-func pipelineE2EStage(id, cmd, needs string) string {
-	block := "  - id: " + id + "\n    cmd: |\n      " + cmd + "\n"
-	if needs != "" {
-		block += "    needs: [" + needs + "]\n"
-	}
-	return block
 }
