+++ internal/cli/pipeline_service_e2e_support_test.go
@@ -1,0 +1,45 @@
+package cli
+
+import (
+	"io"
+	"net/http"
+	"strings"
+	"testing"
+)
+
+// Test helpers shared with UNTAGGED tests, kept out of the e2e build tag
+// (#1760 step 3). Tagging the e2e file beside this one removes it from the
+// default build; these symbols are used by tests that are not e2e, so they
+// would have stopped compiling. Types are moved WITH their methods.
+
+func serviceRequest(t *testing.T, method, url, token, body string) *http.Response {
+	t.Helper()
+	request, err := http.NewRequest(method, url, strings.NewReader(body))
+	if err != nil {
+		t.Fatal(err)
+	}
+	if token != "" {
+		request.Header.Set("Authorization", "Bearer "+token)
+	}
+	request.Header.Set("Content-Type", "application/json")
+	response, err := http.DefaultClient.Do(request)
+	if err != nil {
+		t.Fatal(err)
+	}
+	return response
+}
+
+func readResponse(t *testing.T, response *http.Response) string {
+	t.Helper()
+	return string(readResponseBytes(t, response))
+}
+
+func readResponseBytes(t *testing.T, response *http.Response) []byte {
+	t.Helper()
+	defer response.Body.Close()
+	body, err := io.ReadAll(response.Body)
+	if err != nil {
+		t.Fatal(err)
+	}
+	return body
+}
