+++ internal/cli/task_dismiss_shell_e2e_test.go
@@ -1,3 +1,5 @@
+//go:build e2e
+
 package cli
 
 import (
