+++ internal/workflow/preset_delivery_e2e_test.go
@@ -1,3 +1,5 @@
+//go:build e2e
+
 package workflow
 
 import (
